refactor socket listener function out of Chat and into JetChat;   have all current members send location updates on new-member
